Adiciona um mapa chamado village e algoritmos para carregar e renderizar mapas
