--- assets/village/PokemonLike.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.10" tiledversion="1.10.1" name="PokemonLike" tilewidth="16" tileheight="16" tilecount="1600" columns="40">
- <image source="../../Downloads/PokemonLike.png" width="640" height="640"/>
-</tileset>
