--- assets/village/sem título.tiled-project
@@ -1,12 +1,0 @@
-{
-    "automappingRulesFile": "",
-    "commands": [
-    ],
-    "compatibilityVersion": 1100,
-    "extensionsPath": "extensions",
-    "folders": [
-        "."
-    ],
-    "propertyTypes": [
-    ]
-}
