--- assets/village/sem título.tiled-session
@@ -1,33 +1,0 @@
-{
-    "Map/SizeTest": {
-        "height": 4300,
-        "width": 2
-    },
-    "activeFile": "",
-    "expandedProjectPaths": [
-    ],
-    "fileStates": {
-        "pK.tsx": {
-            "dynamicWrapping": false,
-            "scaleInDock": 1,
-            "scaleInEditor": 1
-        }
-    },
-    "last.imagePath": "C:/Users/Daniel/Downloads",
-    "map.height": 11,
-    "map.tileHeight": 64,
-    "map.tileWidth": 64,
-    "map.width": 20,
-    "openFiles": [
-        "pK.tsx"
-    ],
-    "project": "sem título.tiled-project",
-    "recentFiles": [
-        "pK.tsx"
-    ],
-    "tileset.lastUsedFormat": "tsx",
-    "tileset.tileSize": {
-        "height": 64,
-        "width": 64
-    }
-}
