+++ views/assets/village/CharacterAnimation/Idle/Untitled-0.jpx
@@ -1,0 +1,1 @@
+{"w":16,"h":16,"unit":1,"speed":0.2,"animation":1,"frames":[{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAABC0lEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdechNIhfvBrTJqwG9IYiNMI0g2iiDEDWfPQOAwOID9IIcxW6IRguACkEaQSBBXYMDAmHEM5GloOJYvUCSCPY6QGRDOs2LIebgGwYXgNgkp/6IsFMviKEIejBiNUFII0gTcgGwMQIG1DRxcDQUQbXDNMAdgVUDtkQDBeAA84KYggKqOhiWHCsDCVQQfIoBsCdaezHwHB2E0McgzDYjEUMbxkYoGLoXsFuAAMDXDPMFWBDGBjAXkMOVKyByCBeAla8c4k9w4YNGximrxGEmPOyBz0MUb0Al4UaANeAzkcyBmdmgqVGmFpsqRAjEDHcR4QAACaqaAGTENdEAAAAAElFTkSuQmCC","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAYUlEQVQ4T72T2w4AEAxD6/8/mkgQ6tKJi8fZzroOB8CDThdI944TAcRYkz8rzrUMuQtQ3UcqGgVfAG89iDOqMZYKVPFbE63dWUVZ4zHAYqD0wALZegc5uR5PAmoVDBh95wBfICIBcWavSgAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAABC0lEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdechNIhfvBrTJqwG9IYiNMI0g2iiDEDWfPQOAwOID9IIcxW6IRguACkEaQSBBXYMDAmHEM5GloOJYvUCSCPY6QGRDOs2LIebgGwYXgNgkp/6IsFMviKEIejBiNUFII0gTcgGwMQIG1DRxcDQUQbXDNMAdgVUDtkQDBeAA84KYggKqOhiWHCsDCVQQfIoBsCdaezHwHB2E0McgzDYjEUMbxkYoGLoXsFuAAMDXDPMFWBDGBjAXkMOVKyByCBeAla8c4k9w4YNGximrxGEmPOyBz0MUb0Al4UaANeAzkcyBmdmgqVGmFpsqRAjEDHcR4QAACaqaAGTENdEAAAAAElFTkSuQmCC","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAYUlEQVQ4T72T2w4AEAxD6/8/mkgQ6tKJi8fZzroOB8CDThdI944TAcRYkz8rzrUMuQtQ3UcqGgVfAG89iDOqMZYKVPFbE63dWUVZ4zHAYqD0wALZegc5uR5PAmoVDBh95wBfICIBcWavSgAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA/klEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdechNIhfvBrTJqwG9IYiNMI0g2iiDEDWfPQOAwOID9IIcxW6IRguACkEaQSBBXYMDAmHEM5GloOJYvUCSCPY6QGRDOs2LIebgGwYXgNgkp/6IsFMviKEIejBiNUFII0gTcgGwMQIG1DRxcDQUQbXDNMAdgVUDtkQDBeAA84KYggKqOhiWHCsDCVQQfIoBsCdaezHwHB2E0McgzDYjEUMbxkYoGLoXsFwAUwBTDPMFSBDsIUD1kBkEC8B69u5xJ5hw4YNDNPXCELMedmDkZbxGgDXADWQaAOwpThsYhiBiOE+IgQAJqpoAY96KHwAAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAW0lEQVQ4T71TWwoAIAhb9z90ERToeizI8lPdnLMSgAyKIdHqiRsB1JzrX4E7lkliCdT0mQqn4AvBWw/qjmqNrQIFfmvi6XRWEXvGaxPDCfrJrD/yIVkVTDD7zgVXMCIBJGQ7ZQAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAABCklEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdechNIhfvBrTJqwG9IYiNMI0g2iiDEDWfPQOAwOID9IIcxW6IRguACkEaQSBBXYMDAmHEM5GloOJYvUCSCPY6QGRDOs2LIebgGwYXgNgkp/6IsFMviKEIejBiNUFMI3oirEZhGlARRcDQ0cZA7ohYM1QOWSDMQwAB5wVxBAUUNHFsOBYGUqgguRRDADZCrbJ2I+B4ewmhjgGYbAZixjewsXgaqCmYzeAgQGuGeYKsCEMDGCvIYcF1kBkEC8BK965xJ5hw4YNDNPXCELMedmDHq6oXoDLQg2Aa0DnIxmDMzPBUiNMLbZUiBGIGO4jQgAAPq9iATmmiPAAAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAYUlEQVQ4T72T2w4AEAxD6/8/mkgQ6tKJi8fZzroOB8CDThdI944TAcRYkz8rzrUMuQtQ3UcqGgVfAG89iDOqMZYKVPFbE63dWUVZ4zHAYqD0wALZegc5uR5PAmoVDBh95wBfICIBcWavSgAAAABJRU5ErkJggg=="}],"speed":null}],"palette":[[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0]],"tileset":{"w":10,"h":2,"zoom":2,"tiles":[null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null]},"background":""}
