+++ views/assets/village/CharacterAnimation/WalkDown/Untitled-2.jpx
@@ -1,0 +1,1 @@
+{"w":16,"h":16,"unit":1,"speed":0.15,"animation":1,"frames":[{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA3klEQVQ4T2NkYGD4z0ABYMRmQG8owsR15xkYggwZGED00TuYNmE1wFoFoRCkGQSINgBdM0gjDIAMK16N6goMF8AMgDkbpByfK3B6AdkAmJ1EhwFIA7JXQHxsmkHiWF3wqS+Sga9oOYpnsYlhN6Cii4Gho4wBpAEZgA2EyiGLY7hggR0DQ4IVxBAUUNHFsOBYGUPCIQKxAHcqyDZ00FHGALYAyRAMF6AYsPsIA4OrDcQYEPvsJgwzsQYiWBXIBSBNyIBoA8RLGBhe9qBqxiaGKxphaQBX3OONBcyQwy8CAN4dWAGX/LbMAAAAAElFTkSuQmCC","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAUklEQVQ4T2NkYGD4z4AGMASg8ozoChkYGEBiKOpxaYbpRTeEugYQsh2bK1BcMPAGgJxIyBWjgYhIirCwoH800jYWsKUD9MxC0AXohmDLwsi5GgA7LCEBGuXdzwAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA4klEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcLUdech/OLVmDZhNaA3lIEBpAkGYIYRZQA2zTDDsLkCwwXIBoA0IGsGsY/eQfUGzjAAKUM2AMRH1wwSw2oASAIUkMgAm2acBnzqi2TgK1qOYgA2MewGVHQxMHSUMYA0IAOwgVA5ZHEMLyywY2BIsIIYggIquhgWHCtjSDiEJxBRnAmyDRlADUT3CooLUCSN/RgYXG0gRuw+wsBwdhOYidcAFBtBBiADqAGogriiUbyEgeFlD6pabGK40gEsDcDiHp2PNxbQnUiIDwA6zlwBjyzGdQAAAABJRU5ErkJggg==","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAXUlEQVQ4T92T2w4AEAxD6/8/mkhMKNIJiYTH6Y7uIgCIoDMEyn1gIYAc6/SrZMtlyF2Aen3monPwHpAtKhc/N1HVzktovahjPAZ4JiAXyQPZGqOJ2/IkoHXBgNl3Tl8cIgHhDCs6AAAAAElFTkSuQmCC"}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAzklEQVQ4T2NkYGD4z0ABYMRmgLUKxMQgQ4TJ684zMBy9g2kTTgNAmkGaYADEL15NhAEg27FphhmG7goMF6AbgGwYNlfgDQNYOOCyHSSP1QCQBCwgYb7GFoA4DfjUF8nAV7QcJcSwiWE3oKKLgaGjjAGkARmADYTKIYtjeGGBHQNDghXEEBRQ0cWw4FgZQ8IhVGEUA8CaQQpANqEDqKvQvYY9EI39GBhcbSBG7D4CYZNsALILoAZgyzI4oxFdMShaic4L2DSDxMg2AJvTYWIANdtSAdminSgAAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAXklEQVQ4T2NkYGD4z4AGMASg8ozoChkYGEBiKOpxaYbpRTeEugYQsh2bK1BcMPAGgJxIyBXDORAJ+R05ESKHA3WjkZhYAKnB6QJiDCApGjEyCprtMNfgDT9kSWzZGQAzLCEBc25CYAAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA4klEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcLUdech/OLVmDZhNaA3lIEBpAkGYIYRZQA2zTDDsLkCwwXIBoA0IGsGsY/eQfUGzjAAKUM2AMRH1wwSw2oASAIUkMgAm2acBnzqi2TgK1qOYgA2MewGVHQxMHSUMYA0IAOwgVA5ZHEMLyywY2BIsIIYggIquhgWHCtjSDiEJxBRnAmyDRlADUT3CooLUCSN/RgYXG0gRuw+wsBwdhOYidcAFBtBBiADqAGogriiUbyEgeFlD6pabGK40gEsDcDiHp2PNxbQnUiIDwA6zlwBjyzGdQAAAABJRU5ErkJggg==","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAXUlEQVQ4T92T2w4AEAxD6/8/mkhMKNIJiYTH6Y7uIgCIoDMEyn1gIYAc6/SrZMtlyF2Aen3monPwHpAtKhc/N1HVzktovahjPAZ4JiAXyQPZGqOJ2/IkoHXBgNl3Tl8cIgHhDCs6AAAAAElFTkSuQmCC"}],"speed":null}],"palette":[[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0]],"tileset":{"w":10,"h":2,"zoom":2,"tiles":[null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null]},"background":""}
