+++ views/assets/village/CharacterAnimation/WalkLeft/Untitled-3.jpx
@@ -1,0 +1,1 @@
+{"w":16,"h":16,"unit":1,"speed":0.2,"animation":1,"frames":[{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA70lEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdecR7OLVmDahGNAbysAAUgSiQRrRwdE7RBgAsxVEgzQgGwZyFborMLwAcj7MJhAbBJC9g+4KrGGwwA6iMeEQAwPMEJiL0D2B0wCQZhiAeYNgGMA0fOqLZOArWs4Ao2GxQjAWkJ0H0gwDIMNwAaxewFBc0QUR6ijDkMJpQByDMFjxoopyhriOToZFxtYMDK42GIbgNQCk6VM0N0NO0S7SDcjMTmaYPnUuJCCXfkU4/ewmFG/gDgPxEoTClz0MDMZ+ED7RBoAUwwwBGYADEBcLOLUzMAAABIpaAXe2WhwAAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAYUlEQVQ4T81S2Q4AEAyr//9oMkGYHUgIj9M1PRYARLA3Dcp/4EAANMt4bYnvcJL3BKSoV9EUnNrYJlAzqGF5Yd4nsLIw76Dv27LhWlhpQ63xJMihxn8IdpoQLUhWpAoJlwC2RSABx4ZfbQAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA80lEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdecR7OLVmDahGNAbysAAUgSiQRrRwdE7RBgAsxVEgzQgGwZyFborMLwAcj7MJhAbBJC9g+4KrGGwwI6BIeEQAwPMAJjDCXoBW2SAvADzFkkGfOqLZOArWg53BTb/gwzG6gWQBMgAGAAZhAvgNABFQ0UXhNtRhmEOTgPiGITBihdVlDPEdXRC2MbWDAxnN6EYgmEAzO9wA4ytGT5FczPkFO0izgCY8ZnZyQzTp86F2wY2eOlXwi5gEC9B9efLHgjf2A9CE/ICWBHIEJhGbHwkK4iLBZyRyMAAAM8+WgFBvxXxAAAAAElFTkSuQmCC","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAWklEQVQ4T2NkYGD4z4AGMASg8ozoChkYGEBiYPW4NKHrQTeE/gaAXITsCrgLyPUGyQbgDANYYBEKTNobgC8s8KYD5PjG5w2KvUDbaCQ2LQyRhITsTFisYItGAAIsIQEl8kHeAAAAAElFTkSuQmCC"}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA70lEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdecR7OLVmDahGNAbysAAUgSiQRrRwdE7RBgAsxVEgzQgGwZyFborMLwAcj7MJhAbBJC9g+4KrGGwwA6iMeEQAwPMEJiL0D2B0wCQZhiAeYNgGMA0fOqLZOArWs4Ao2GxQjAWkJ0H0gwDIMNwAaxewFBc0QUR6ijDkMJpQByDMFjxoopyhriOToZFxtYMDK42GIbgNQCk6VM0N0NO0S7SDcjMTmaYPnUuJCCXfkU4/ewmFG/gDgPxEoTClz0MDMZ+ED7RBoAUwwwBGYADEBcLOLUzMAAABIpaAXe2WhwAAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAYUlEQVQ4T81S2Q4AEAyr//9oMkGYHUgIj9M1PRYARLA3Dcp/4EAANMt4bYnvcJL3BKSoV9EUnNrYJlAzqGF5Yd4nsLIw76Dv27LhWlhpQ63xJMihxn8IdpoQLUhWpAoJlwC2RSABx4ZfbQAAAABJRU5ErkJggg=="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA6ElEQVQ4T2NkYGD4z0ABYMRmgLUKA0OQIcTUdecR7OLVmDahGNAbysAAUgSiQRrRwdE7RBgAsxVEgzQgGwZyFborMLwAcj7MJhAbBJC9g+4KrGGwwA7h1IRDDAwwgwh6AdmHIENAmmEA2WXI6rC6AKTgU18kA1/RcjgNC2D0YMRrAEwxyCBcAKcBKBoquiDcjjIMc3AaEMcgDFa8qKKcgWH3EQYGVxushuA1YEqfG0NO0S6GRcbWpBsAC0QGYz9UZ5/dhMLHdIF4CQPDyx6IIhAbBEB8mEEEDYBpxGYIlqggLhZwRiIDAwAsplcB5SjFhAAAAABJRU5ErkJggg==","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAYElEQVQ4T9VS7Q4AEAi83v+hGcMShTY2fqau+4gABIg3FMo/yUYAqZb7tSE5I0HeAyRGnEVj4JVxDKB6UM1amXkfwPLCvAOetyVjKWEnDTVGj5FdjDvbPzqkahRPZBZjBPolIQFYWnsZAAAAAElFTkSuQmCC"}],"speed":null}],"palette":[[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0]],"tileset":{"w":10,"h":2,"zoom":2,"tiles":[null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null]},"background":""}
