+++ views/assets/village/CharacterAnimation/WalkRight/Untitled-3.jpx
@@ -1,0 +1,1 @@
+{"w":16,"h":16,"unit":1,"speed":0.2,"animation":1,"frames":[{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA7ElEQVQ4T2NkYGD4z0ABYMRmQG8oxMR15xkYggwR7KN3MG3CaoC1CqZCkEHFqxkYQIaDaBjAMACkAGQzCMA0wQyEuQavASDFyM4GGQRzOkgO3RsEvQDSsMAO4qKEQySEAcz5MC0gQ4g2ABYOINs/9UUy8BUth9PobsDqBZgikGYYABmCDWA3oKILorajjGASwzQApHn3EYa4s0cZFlWUM8R1dIINWcTwlkgXIBkwpc+NgW/pV4hhRBtg7Ae36VM0NzgAM7OTGaZPnUukC2AGnN3EwCBegtD0sodIA5CVwQzAoRmkFG80EowCoAIA7IpdARQJNr0AAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAXElEQVQ4T2NkYGD4z4AGMASg8ozoChkYGEBiKOpxaUbXCzOMugYQazvINVhdQKwByGFBlhcoMgA9Jkh2AUEDQAFEs3RAMCERioWRYACpMQBL0ljDDlkQW+DBcicAzkUgAWYvCU4AAAAASUVORK5CYII="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA5UlEQVQ4T2NkYGD4z0ABYMRmQG8oxMR15xkYggwR7KN3MG3CaoC1CqZCkEHFqxkYQIaDaBjAMACkAGQzCMA0wQyEuQavASDFyM4GGQRzOkgO3RtEeQGkaYEdA0PCIRLCAJtzsUUWzliAhQPI9k99kQx8RcuxRjZWA2AqQRphgDQDKrog+jrKCCYxTBcY+zHEnT0K1rioopwhrqMTwmZ4S6QXoAZM6XNj4Fv6FWEYw1usYYHVBZ+iuVECLTM7mWH61LnEuwCs8uwmiAbxElSNL3tQ+LhjAaQRWTE6H2oM3mgkGAVABQCXAlkBAhje9QAAAABJRU5ErkJggg==","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAa0lEQVQ4T7WTUQ6AIAxDy/0PLVniTK06KgpfBLZHu40GYIOsy8F+3zQQQJyd4p+SNTdh/wLc10PNrQIXwLWYsvAJoJ14rWAIiAItm4PhIDldKC04AJ6B3B95DmCtgqoD/JnKQWIIB6Y9tdAB6h0hAbQLqN8AAAAASUVORK5CYII="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA7ElEQVQ4T2NkYGD4z0ABYMRmQG8oxMR15xkYggwR7KN3MG3CaoC1CqZCkEHFqxkYQIaDaBjAMACkAGQzCMA0wQyEuQavASDFyM4GGQRzOkgO3RsEvQDSsMAO4qKEQySEAcz5MC0gQ4g2ABYOINs/9UUy8BUth9PobsDqBZgikGYYABmCDWA3oKILorajjGASwzQApHn3EYa4s0cZFlWUM8R1dIINWcTwlkgXIBkwpc+NgW/pV4hhRBtg7Ae36VM0NzgAM7OTGaZPnUukC2AGnN3EwCBegtD0sodIA5CVwQzAoRmkFG80EowCoAIA7IpdARQJNr0AAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAXElEQVQ4T2NkYGD4z4AGMASg8ozoChkYGEBiKOpxaUbXCzOMugYQazvINVhdQKwByGFBlhcoMgA9Jkh2AUEDQAFEs3RAMCERioWRYACpMQBL0ljDDlkQW+DBcicAzkUgAWYvCU4AAAAASUVORK5CYII="}],"speed":null},{"layers":[{"name":"Layer #1","visible":true,"color":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAA4ElEQVQ4T2NkYGD4z0ABYMRmQG8oxMR15xkYggwR7KN3MG3CaoC1CqZCkEHFqxkYQIaDaBjAMACkAGQzCMA0wQyEuQavASDFyM4GGQRzOkgO3Rt4vQBSvMAO4dyEQySEAbJNIEOwaQYZhzMWQP781BfJwFe0HE5ji22sBsAUggyAAZBBxBtQ0QVR21FGMIlhugCmefcRBgZXG4a4jk6wIYsY3hLpAiQD4s4eZZjS58aQU7SLBAOM/VBtOruJxECEGXB2EwODeAnEsJc9EBrEh7Gh1uCOBSI040wHBIMeSQEAGZdZAb6s4dwAAAAASUVORK5CYII=","alpha":"data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAAaUlEQVQ4T8WTSQ4AIAgD6/8frSGRBOuCa/SIMBaoAUAEnSqQ7wMnApBYkd8r5lqF3QXMvi5qmgpmAXYWWy0cAXgTywpcgAzomQ9cI3lb+A+wBlJrF1vwWrBFTcBoA/YzDY3EEE1WdTzIBOolIQEcGbbBAAAAAElFTkSuQmCC"}],"speed":null}],"palette":[[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0],[255,255,255,0]],"tileset":{"w":10,"h":2,"zoom":2,"tiles":[null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null,null]},"background":""}
